mm: Export do_madvise()
Currently, do_madvise() is only used by madvise() system call
to allow kernel to make better use of memory for applications
in userland.

It could also be used in kernel internally. E.g. KVM can use
it to provide guest page hints to VM management subsystem in
host. In order to solve this lack of support, this patch exports
do_madvise() so as to make it available to the entire kernel.
The already existing madvise() system call is unchanged and
now relies on this new do_madvise() function.

Suggested-by: Arjan van de Ven <arjan.van.de.ven@intel.com>
Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>
Signed-off-by: Kemi Wang <kemi.wang@intel.com>