x86: kvm: Return memory from guest to host proactively
To use the system memory efficiently, the memory size of guest should
be dynamically adjusted. The design principle is quite straightforward.
When a free page is added into buddy system of a guest, it indicates that
this page is unused in the guest and a hypercall is called to mark it
lazyfree on host. Thus, when the host is under memory pressure, page reclaim is
invoked and the page which was marked lazyfree is released to host.

Implementation details:
A new hypercall KVM_HC_RETURN_MEM is introduced, relying on the do_madvise()
exported before, allows the guest to provide guest memory hints to host. This
hypercall is called in the page *free* codepath of page allocator in guest
when the page order is
no less than 7. This value has been found running memory tests
using multiple threads allocating/freeing high amount of memory.
Those tests were run for different order values, and 7 was the best
tradeoff between the number of hypercalls issued and the amount of
memory returned to the host.

Addtionally, the MADV_FREE advice is used by do_madvise() as default to
mark pages lazyfree on host. Thus, page reclaim on host will release
these pages back to its buddy system under memory pressure.
Also, if this page is revisited on the guest before returning back to
host buddy system, since the page is still in process address space
with page table mapping(MADV_FREE advice), it will not trigger a new
page fault, page allocation and zero on host. Thus, it will not degrade
guest performance.

Optimization tricks:
a) There may be memory fragmentation in guest memory when a
guest has run for a long time. Then, high order pages in guest may not
be as many as before, leading to low efficiency of our approach. To
overcome this drawback, the userspace daemon *memshinker* can
triggers a compaction round once every hour or two.
https://github.com/clearlinux-pkgs/memshrinker

b) Also, page caches may eat up the guest memory when a guest has run
for some time, and most of them are rarely revisited. To drain these
page cache pages, an extra sysctl interface is introduced to release these
pages back to guest buddy system (Need explain why we need extra sysctl
insted of drop_caches).

Suggested-by: Arjan van de Ven <arjan.van.de.ven@intel.com>
Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>
Signed-off-by: Kemi Wang <kemi.wang@intel.com>